# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/blog/posts/search": {
      "get": {
        "operationId": "BlogController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "q must be at least 2 characters"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/posts": {
      "get": {
        "operationId": "BlogController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      },
      "post": {
        "operationId": "BlogController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 409
                    },
                    "error": {
                      "type": "string",
                      "example": "Conflict"
                    },
                    "message": {
                      "type": "string",
                      "example": "Slug already exists"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/posts/recent": {
      "get": {
        "operationId": "BlogController_getRecentPosts",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 10,
              "default": 5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/categories": {
      "get": {
        "operationId": "BlogController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      },
      "post": {
        "operationId": "BlogController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/categories/{id}": {
      "put": {
        "operationId": "BlogController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      },
      "delete": {
        "operationId": "BlogController_deleteCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/tags": {
      "get": {
        "operationId": "BlogController_getTags",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 15,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/posts/{slug}": {
      "get": {
        "operationId": "BlogController_findOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Post not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      },
      "put": {
        "operationId": "BlogController_update",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Post not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      },
      "delete": {
        "operationId": "BlogController_remove",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Post not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/posts/{slug}/preview": {
      "get": {
        "operationId": "BlogController_findOnePreview",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Post not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/posts/{slug}/related": {
      "get": {
        "operationId": "BlogController_getRelatedPosts",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Post not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
    "/api/blog/images": {
      "post": {
        "operationId": "BlogController_uploadImage",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "No file provided or invalid file type"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "blog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hyunwoo API",
    "description": "chahyunwoo.dev blog & portfolio API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "PostTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "React"
          },
          "slug": {
            "type": "string",
            "example": "react"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "PostSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Next.js 15 App Router 완전 정복"
          },
          "slug": {
            "type": "string",
            "example": "nextjs-15-app-router"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "미입력 시 content에서 자동 추출"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://assets.chahyunwoo.dev/thumbnail/example.png"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "example": "Frontend"
          },
          "published": {
            "type": "boolean",
            "example": true
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "readingTime": {
            "type": "number",
            "example": 5,
            "description": "분 단위 예상 읽기 시간"
          },
          "viewCount": {
            "type": "number",
            "example": 128
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostTagDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "description",
          "thumbnailUrl",
          "category",
          "published",
          "publishedAt",
          "readingTime",
          "viewCount",
          "createdAt",
          "updatedAt",
          "tags"
        ]
      },
      "PostListResponseDto": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostSummaryDto"
            }
          },
          "total": {
            "type": "number",
            "example": 40,
            "description": "필터 조건에 맞는 전체 건수"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "posts",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Frontend"
          },
          "icon": {
            "type": "string",
            "example": "Monitor",
            "description": "lucide-react 아이콘 이름"
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Next.js 15 App Router 완전 정복"
          },
          "description": {
            "type": "string",
            "description": "미입력 시 content에서 자동 추출"
          },
          "content": {
            "type": "string",
            "description": "MDX content"
          },
          "category": {
            "type": "string",
            "example": "Frontend"
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "이미지 업로드 API에서 받은 URL"
          },
          "published": {
            "type": "boolean",
            "default": false
          },
          "publishedAt": {
            "type": "string",
            "example": "2026-03-21",
            "description": "발행일 (미입력 시 발행 시점 자동)"
          },
          "tags": {
            "example": [
              "React",
              "TypeScript"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Next.js 15 App Router 완전 정복"
          },
          "description": {
            "type": "string",
            "description": "미입력 시 content에서 자동 추출"
          },
          "content": {
            "type": "string",
            "description": "MDX content"
          },
          "category": {
            "type": "string",
            "example": "Frontend"
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "이미지 업로드 API에서 받은 URL"
          },
          "published": {
            "type": "boolean",
            "default": false
          },
          "publishedAt": {
            "type": "string",
            "example": "2026-03-21",
            "description": "발행일 (미입력 시 발행 시점 자동)"
          },
          "tags": {
            "example": [
              "React",
              "TypeScript"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}